Fix story code rendering
The "show code" option no longer works with a story function declared
like
```
export const () => <...>
```
It requires that an `args` is passed, like
```
export const args => <...>
```
(note the args don't have to actually be used).

Alternatively we can set `.parameters.docs.source.code` to `null` for
the story, which disables the "show code" option entirely.

Chances are these could be improved a bit, to declare proper types for
the args and so on. I'll leave that to people who know what these
components are supposed to be doing.