Gardening: don't trigger automated triage during triage
The Triage Issues task runs when new issues are being created, but also when issues are labeled.
We do not want to start adding labels automatically (doing our own automated triage) when the action that triggered this was someone making changes to labels that would impact our automated triage.

So our automated triage must be aware of labels being added to the issue right now, in addition of being aware of issues already existing on the issue.

See p1681411170005189/1681404941.879739-slack-C04TRTZUSP9