Remove builds on `pnpm install`
It's annoying to have to build these on `pnpm install`, and slows down
the install unnecessarily if the built packages aren't going to be used.
Let's fix that.

To keep the actual npmjs packages usable, we still need external users
to get the built versions while getting the sources within the monorepo.
package.json exports custom conditions are a way to do that: we specify
a custom condition "jetpack:src" first pointing at the source index
file, then "types" for the index.d.ts where appropriate, then "default"
for the built index.js.

Then we have to configure tooling inside the monorepo to look for
that "jetpack:src" condition:

* Webpack uses `.resolve.conditionNames`. But in case we ever publish
  our webpack-config package, let's avoid the problem we had with
  calypso-build and their "calypso:src" condition by setting an
  environment variable via `.npmrc` and using that instead of
  hard-coding 'jetpack:src'.
  * Unlike Calypso, we're not going to also set `.resolve.mainFields`.
    Exports are the modern way to do things like this.
* Typescript's tsc uses `.compilerOptions.customConditions`. Plus we also
  have to set `.compilerOptions.moduleResolution` to "bundler" for it to
  pay attention to exports at all (or "nodenext", but that has a lot of
  other side effects).
* Jest (at least for the jsdom and node environments) uses
  `.testEnvironmentOptions.customExportConditions`.
* Eslint's `eslint-plugin-import` needs a custom resolver (sigh).
* Rollup uses `exportConditions` in the options for
  `@rollup/plugin-node-resolve`.

Then, for good measure, let's also have the cli strip the "jetpack:src"
from the mirror repos to make it even less likely that some external
user gets confused by it.