Bump PHP version we test from 7.2 to 7.3
It is not super ideal but PHPUnit 9+ requires PHP 7.3 or newer.
We could, technically, try and downgrade some of our packages but we have a lot of conflicts with:

1. We need to use PHPUnit 8 or older for PHP 7.2 to be supported.
2. Many of our dependencies only support PHPUnit 7 or 9 (so not 8) which means we're limited to 7.
3. We have other dependencies that doesn't go as far back as PHPUnit 7.

In the end then I truly believe that running our tests on PHP 7.3+ environments (even though we list 7.2+ support) will - in
the end - give us much better/safer code since the changes between the two PHP versions are minimal but it opens up for us
adding unit and integration tests which should generally.
We also have linting that checks for 7.2 compatible code, so we should not be able to introduce breaking code anyway.