[not verified] ActivityPub: add compatibility functions
The ActivityPub plugin returns an object with information about each post when you make a ActivityPub request. This includes an image for the post, fetched using core's wp_get_attachment_image_src.

Jetpack's Image CDN changes the URL of the images returned by that function. In some scenarios, that may not be desirable (Some Fediverse instances may not always be able to fetch images from the Jetpack CDN, for example).

In this commit, we add a new compatibility layer for the ActivityPub plugin, where we allow disabling Jetpack's Image CDN processing when fetching an image for ActivityPub. This is only enabled if you use a filter.

Discussions:

- https://github.com/pfefferle/wordpress-activitypub/pull/307
- https://github.com/pfefferle/wordpress-activitypub/pull/309

Internal reference: p1682284516722569-slack-C04TJ8P900J