Repo Gardening: automatically add legacy Photon label
Photon now lives on in the Image CDN package. We can use the new "[Package] Image CDN" label to look for changes to the feature, but the existing Photon label is still widely used. Let's add that legacy label to PRs that touch the package, so we can continue to use the Photon label for sorting through issues and PRs in the repo.