beta plugin: Fix markdown rendering (#30339)
The markdown library we're using seems to get confused by HTML-style
tags and might let some through that shouldn't be let through. Since it
has a flag to disable HTML-style tags entirely, and we don't use those
much in PR summaries or the relevant bundled .md files, let's set that
flag to avoid the problem.