Remove `@types/wordpress__components`
GB switched to including types in the package instead of using a
`@types/` package.

This also necessitated some code changes, since they're not publishing
the component props (see WordPress/gutenberg#50157).

Also of note is that their types for `ClipboardButton` are broken, in
that they didn't declare that it also accepts all `Button`'s props.
Since that component is deprecated anyway, we just replace use of it
with the designated replacement.