tools: Add list-repo-maintenance-commits.sh
We want a "monorepo maintenance" project thread (e.g. p9dueE-77C-p2) so
managers can have something to look at to see the maintenance stuff we
do. This script helps filter the commits for a week to ones that have
some likelihood of being "maintenance" to make it easier to list them.