Social: Update the UI logic for checking broken social connections
With the move to using the connection ID rather than the token ID we
needed to update the logic in the classic editor so that connections are
correctly identified as broken.

Similarly, there has been a longstanding problem with the twitter
options section of the publicize-form where it only checks the `enabled`
property. This needs to also check the `is_healthy` property to also
disable the options if there are no 'healthy' twitter connections.