Add: WPCOM API request tracker to debug helper
Some recent Atomic reverts were due to unnecessary-large numbers of
WPCOM API requests. This change aims to add visibility to these requests
to make it easier to catch problems related to the number of requests
before shipping the code and experiencing problems in production.

With the new WPCOM API Request Tracker module activated in the Debug
Tools, an admin bar menu entry "WPCOM API Requests" appears. After a
page loads, the entry title updates to indicate the total number of
requests. Clicking on the entry toggles the display of a panel with
further details.