ci: When checking changelogger use, count the changelog itself
Our "Check changelogger use" check looks for commits that change a
project without adding a change entry. This breaks in the edge case
where we need to release a new build of a project that doesn't have any
existing change entry files: the change to the changelog itself is
ignored, but any changes to other files for version bumps will trigger
the check.

Instead of ignoring the change to CHANGELOG.md, let's instead count it
the same as having a change entry being touched.