Release packages/blaze (#30850)
It seems that the redux store in js-packages/connection is global rather
than per-bundle, so when something is added to that then all the deps
that bundle it should get rebuilt.

Looks like blaze is the only package that didn't get a release already,
and 🤞 the standalone plugins that bundle it directly (not via packages)
look like they all either only load it on their own admin pages or check
for Jetpack first.

I think this will cover Jetpack-the-plugin along with any of our
standalones too. Not so sure about a site with say Social and VideoPress
but not Jetpack-the-plugin; hopefully a real fix gets in before either
of those gets a release.