tools: Fix deploy-to-svn.sh stable tag handling (#31114)
When we update the stable tag in w.org SVN, we usually forget to bring
that back to the monorepo. This can then too easily result in the stable
tag getting accidentally downgraded on a future deploy.

We already handle this correctly in the wp-svn-autopublish action, let's
use the same logic here to ensure that the tag in `trunk/readme.txt` is
never changed by the script.