Pass filter status and site values from PHP to JS
We'll use those in the editor panel, instead of relying on data available via Jetpack_Editor_Initial_State. This is better for us on the long run, since we want our editor panel to work on sites where the Jetpack plugin isn't installed, and where Jetpack_Editor_Initial_State consequently will not be available.