autoloader: Store own version in autoload data
If you do a local build, the version stored in the autoloader for all
monorepo packages will probably be "dev-trunk" rather than the actual
version of the package.

For the most part we can't really do anything about this. But for the
autoloader itself, we can have it store its own version, which can avoid
certain bugs when trying to rsync to a wpcom sandbox for testing
purposes.

Also, to further avoid this kind of bug, let's include the autoloader's
own version in the "random" namespace used for the copied classes.