Markdown Block: render single and double quotes as smart quotes
Fixes #19549

This involves enabling a parameter in markdown-it: "typographer".
The parameter is described as:

> Enable some language-neutral replacement + quotes beautification
> For the full list of replacements, see https://github.com/markdown-it/markdown-it/blob/master/lib/rules_core/replacements.js

It seems safe enough to enable in Jetpack's Markdown block. It also brings in replacements that are very close (with a few exceptions) to what you get out of the box with WordPress, with Core's Texturize:
https://developer.wordpress.org/reference/functions/wptexturize/
It consequently seems like a good idea to implement this default behavior in the Markdown block.