[not verified] Jetpack: Fix newsletter editor sidebar panel in WP 6.1 (#31386)
PR #30958 added some code to return early from the sidebar panels when
the subscriptions module is disabled. Unfortunately it did this in a way
that only works with React 18, and WordPress 6.1 (which we still
support) comes with React 17.

Fixing this for React 17 is simple enough, we just have to return `null`
instead of implicit `undefined` in the early return.