ci: Checkout package trees for build
In #31220 we added logic to the build to append a timestamp to package
alpha versions, derived from the last git commit to the package.

But in CI we use a shallow checkout, so the "last git commit" is the
grafted commit. To get the correct timestamp we need to deepen the
checkout until `git log -1` returns a non-grafted commit for each
package that has any change entries.

On the plus side, we can use `--filter=blob:none` to fetch only the
directory tree metadata and not all the file contents.