Publicize Components: Refactor the media validation
We need to change the way we validate the media to be shared. Previously
we checked the lowest requirements for all enabled connections, but
given the way that Instagram requires an image we were forcing the media
to be valid for Instagram, which is currently only a JPG.

This change reworks the validation so we do it on a per connection basis
and disable the individual connections if the selected media is invalid.