phpcs: DRY in .phpcs.dir.phpcompatibility.xml (#31573)
For cases where we want to ignore a PHPCompatibility rule, we need to
ignore it in both `.phpcs.dir.xml` and `.phpcs.dir.phpcompatibility.xml`.

But there's no need to do it via copy-paste. We can have
`.phpcs.dir.xml` load `.phpcs.dir.phpcompatibility.xml` so we only have
to have the ignores in one place.