Update postcss-loader configs
A few years ago postcss-loader added the ability to specify plugins as
an array of plugin objects rather than as an object mapping plugin names
to configs.

Taking advantage of this lets us not have to use `NODE_PATH` for webpack
invocations all over the place to allow postcss-loader to find its
plugins.