gardening: Remove `[Status] Stale` from closed issues/PRs
We're trying to better deal with the backlog of stale issues and PRs. It
seems like it would help with that to remove the Stale label from them
when they're closed.

This also means we have to start running the cleanLabels task when
issues are closed. Fortunately we already run the action on issue close
for other reasons, so nothing needs to change in the workflow itself.