codesniffer: Add "compat" rulesets for per-dir configs (#31711)
One annoying thing about the per-dir configs is that we can't set
PHPCompatibility's `testVersion` per directory. We work around that by
instead excluding the relevant rules in the config.

Unfortunately it's easy for people to screw this up, e.g. ignoring all
of `PHPCompatibility.FunctionDeclarations.NewReturnTypeDeclarations`
when some like `.mixedFound` shouldn't be.

So let's make some centralized rulesets that ignore only the rules that
need ignoring per PHP version so they can be used in the per-directory
configs. It looks like PHPCompatibility has pretty comprehensive tests
based on fixture files, so we should be able to get a pretty
comprehensive list of the rules by running phpcs over those files.