Fix broken type declaration
Apparently the return types of `setTimeout()` and `setInterval()` depend
on the environment, in browsers they're numbers while in node they're
`Timeout` and `Timer`. I have no idea why, but some update here
(assocated with the "expect" package?) makes tsc suddenly start
complaining about this.

Some quick searching suggests that the way to do this is to either
remove the type declarations or do like `ReturnType< typeof setTimeout >`
to let typescript figure it out. I went with the latter, since
someone felt that declaring the type here in the first place was a good
idea.

But to make *that* work, we also need to rename the `ReturnType` type
variable to something else so it doesn't mask TypeScript's built-in
`ReturnType<>` utility type. Whee.