Social Links: Prevent their use in block themes.
We want to make sure that the customizer is not being used for block
themes. This change updates the check before initialising the Social
Links feature, to make sure that a block theme is not active.

This shouldn't be a problem as block themes are unlikely to declare
supoort for Social Links, but in case another plugin does, this ensures
the feature isn't enabled.