crm: Remove composer `.config.platform.php` setting (#31781)
This configuration setting is a blunt instrument, causing composer to
install dependencies for a specific version of PHP regardless of what is
actually being run. It looks like it was added at some point
pre-monorepo when they were bundling a bunch of oauth libs, before
(still pre-monorepo) they switched to downloading them at runtime. As
things currently stand, this doesn't seem to make any difference to the
composer packages that do get bundled.

On the other hand, it does cause some issues in our CI: versions of
symfony and codeception compatible with 7.2 are brought in, which raise
deprecation warnings with 8.1 and 8.2.

Since it doesn't seem to be needed anymore, let's drop it. We already
have some checks in CI that will flag if the non-dev deps accidentally
drop support for any version of PHP that we run tests with.

In fact, we also need to update those checks to not run if the plugin
doesn't support the PHP version.