Jetpack: Replace some bad `plugin_basename()` calls (#31997)
WordPress's `plugin_basename()` function is supposed to be passed a full
path (e.g. `__FILE__`), which it transforms if possible into a path
relative to `WP_PLUGIN_DIR` or `WPMU_PLUGIN_DIR`.

We have a few places that are not using it for that purpose. Instead
they're passing a relative path, which will normally just return the
input but occasionally in the middle of Atomic updates can return a
completely incorrect path. What we really want to do in these places is
append the relative path to `__DIR__`.