Blaze: always display menu in Calypso for Jetpack sites
This is a follow-up to #31617 and #31750.

We now have Blaze::enable_blaze_menu() to add the Blaze menu to wp-admin whenever necessary. However, Jetpack sites still rely on the admin-menu endpoint to display a list of custom menu items in Calypso. There, we'll want to always display the Tools > Advertising menu item, as long as the site is eligible for Blaze.