SSO: enable 'Match by Email' by default (#32370)
This PR changes the default behavior of "Match by Email" functionality. This option has historically been enabled by default since the beginning of times. In #27102 we made the default be off for consistency with its representation in the UI, and it stayed like this for the last 8 months. But there were consequences. See p9GBOq-5Re-p2. With this update, the option will be reactivated by default for all users who have never modified this particular setting.

Before: Match by Email disabled by default
After: Match by Email is enabled by default, and only disabled if an admin manually disables it.