Repo Gardening: label issues with support references
This is something that has happened via a GitHub webhook for the past few years.
However, since we now have a task that does look for support references in issues, we may as well add the label through that task and retire the webhook imo.

Internal reference: p1691543440826329-slack-CQD1HH4MA