repo-gardening: Update docs for new `add_labels` feature
The inline comment and added doc didn't correctly represent the
structure of the input (although the added example did).

The example can use yaml's multi-line string syntax to be represented in
a clearer manner.

May as well take the opportunity to be a little more explicit in the doc
too.