Fix a Global Styles UI performance issue
Instead of loading the Jetpack blocks CSS for the Site Editor using the admin_init hook, we'll switch to the `enqueue_block_editor_assets` hook. In this manner, the CSS script is loaded in the iframe, instead of the main site-editor.php page.

By doing so, the UI for Global Styles is more responsive.