ci: On the prerelease branches, don't build/test plugins
When we're running CI builds and tests on pushes to the prerelease
branche, we're mainly doing that to build Composer packages for use in
an upcoming plugin release branch. There's no point in building the
plugins as they'll be released later when the release branch is created
instead. So let's only consider projects without a release-branch-prefix
set as being "changed".

There is a slight risk that something in the prerelease branch might
break a plugin build, which we'd now not discover until after the
package releases are pushed to packagist and the plugin release branch
is created. But that's probably ok, as the intent of the prerelease
branch is that the only difference between it and trunk (or an existing
release branch if we're doing a point release) are changelog file
updates and version bump, which should not cause that sort of problem.