Social Menu & Social Media Icons: Add X support
Update the icon font, the social icons widget, the social menu, the social-logos package to include the new X icon.

Noting that this purposefully does not replace the existing Twitter icon. This will allow site owners to decide whether the icon they'd like to use. It should be less forceful that way, we will not update the existing Twitter icon that may be in use on many sites; site owners will do that themselves if they'd like.

Related PR: https://github.com/Automattic/social-logos/pull/126/