tests: Replace deprecated PHPUnit `assertObject(Not)HasAttribute`
PHPUnit 10 at first removed this but then in 10.1 re-added it as
`assertObject(Not)HasProperty`. Now that a replacement exists,
yoast/phpunit-polyfills provides a polyfill, so let's make use of it.

Note WordPress's `WP_UnitTestCase` and WorDBless's
`WorDBless\BaseTestCase` both already include all Yoast polyfill traits,
so explicit use of the trait is only needed for things using PHPUnit's
base class.