renovate: Work around renovate bug
Renovate has [a bug][1] where they modify `.npmrc` and don't clean up after
themselves, resulting in those modifications being included in the diff
when post-upgrade commands are used.

Further, it seems they're reluctant to even admit this is actually a
bug, and would rather cast aspersions than collaborate on a fix.

Work around it by manually reverting the file in the post-upgrade script.

[1]: https://href.li?https://github.com/renovatebot/renovate/issues/23528