docker: Add a `--php` option to `jetpack docker phpunit` (#32979)
To make it easier to run tests locally with a specified version of PHP.

This turns out to be a bit more complex than it sounds, as we need to be
able to install the necessary version of PHP inside the container and we
need to make sure the correct versions of Composer deps (particularly
phpunit) are installed too. And then use those rather than whatever is
in the monorepo install.