i18n-check-webpack-plugin: Provide a way to not break with webpack's mangleExports
Webpack's optimization.mangleExports option, enabled by default in
production mode, can on occasion mangle an export to the name of one of
the i18n functions this module looks for.

To fix this, we provide a slightly modified copy of Webpack's internal
`MangleExportsPlugin` that will never mangle to `__`, `_n`, `_x`, or `_nx`.

Then update our webpack-config package to disable `mangleExports` in
favor of this new plugin.