phpcs-filter: Ignore `--basepath` option
I had thought the use of this would be to point to the base of the
monorepo. But it seems likely IDEs like VSCode are using it to point to
various subdirectories of the monorepo instead.

Since, on re-reading the help docs for the parameter, that does seem
like a valid use for the parameter, we shouldn't try to use it to
determine the base of the monorepo after all.