Update PHP syntax in tests
Codeception v5 needs "protected array $requiredFields = ..." to set it
as a field, but PHP 7.3 (which is lowest version CRM supports) doesn't
support that syntax. Looks like we can set it in the `_beforeSuite`
function instead.

Also one function now needs a `void` return type in v5. Fortunately
declaring that is compatible with v4 and PHP 7.3, so we can just do it.