Connection Toggle: Prevent onChange when the component is disabled
When the toggle was disabled, such as when the site had no shares
remaining, or if the post didn't have an image and wanted to share to
Instagram, the `onChange` handler was still being called when the toggle was
clicked.

This change prevents that by checking the state of the `disabled` prop
first before calling the event handler.