Connection: add partner codes are added to connection URLs
This was previously only done in the Jetpack plugin, in 2 ways:

- The Partner package is initialized on plugins_loaded in the main Jetpack class.
- In Jetpack's `build_authorize_url` method we have a `jetpack_build_authorize_url` filter that is leveraged by the Partner package.

Standalone plugins, on the other hand, do not use `Jetpack::build_authorize_url()`. This method is obviously only available in the Jetpack plugin, and so standalone plugins do not have that `jetpack_build_authorize_url` filter.

We fix this discrepancy in this commit:

- We now initialize the Partner package in the connection package.
- We add the `jetpack_build_authorize_url` filter to the connection package's `get_authorization_url` method, so the partner codes can be added to all URLs.

**Note**: this does not add the Partner package as a new dependency for the Connection package, since the Partner package already declares the connection package as a dependency.