webpack-config: Default to 'source-map' in dev mode
It seems our developers have spoken: 25 webpack configs in the monorepo
have been setting webpack's `devtool` to 'source-map' for development
builds, while only 8 have used the default from webpack-config (and 2
are explicitly setting no devtool).

So let's change the default to 'source-map' so everything can be working
the same way.

Note this doesn't change anything for production builds, this is only
for development.