ci: Lint for projects having both tsconfig and jsconfig (#33746)
The files `jsconfig.json` and `tsconfig.json` serve roughtly the same
purpose, with the former being for JS-only projects and the latter for
projects using TypeScript. Having both is unnecessary and confusing, so
lint for that.

In the future we might add lints for when one or the other is needed,
but first we have to determine when exactly that is.