Add subscriber-related tasks to launchpad
New tasks related to subscribers have been added to the launchpad. The tasks added are 'import_subscribers', 'share_site' and 'add_subscribe_block' which allow users to import existing subscribers, share site and add subscribe block to the theme respectively. Logic for checking if the subscribe block has been added to the template part content has also been included, marking the task as complete when the block is found.