Subscriber page: add Launchpad tasks & task list (#33948)
* Add subscriber-related tasks to launchpad

New tasks related to subscribers have been added to the launchpad. The tasks added are 'import_subscribers', 'share_site' and 'add_subscribe_block' which allow users to import existing subscribers, share site and add subscribe block to the theme respectively. Logic for checking if the subscribe block has been added to the template part content has also been included, marking the task as complete when the block is found.

* Add subscriber-related tasks to launchpad

New tasks related to subscribers have been added to the launchpad. The tasks added are 'import_subscribers', 'share_site' and 'add_subscribe_block' which allow users to import existing subscribers, share site and add subscribe block to the theme respectively. Logic for checking if the subscribe block has been added to the template part content has also been included, marking the task as complete when the block is found.

* Extend subscribe block check to all saved templates

* Add check for FSE

* Update projects/packages/jetpack-mu-wpcom/src/features/launchpad/launchpad-task-definitions.php

Co-authored-by: daledupreez <dale@automattic.com>

* Change wording

* Update projects/packages/jetpack-mu-wpcom/src/features/launchpad/launchpad.php

Co-authored-by: daledupreez <dale@automattic.com>

* Rebase

* Bump version of jetpack-mu-wpcom and mu-wpcom-plugin

* Update versions

* Update projects/packages/jetpack-mu-wpcom/changelog/add-subscriber-launchpad-tasks

Co-authored-by: Ivan Ottinger <25105483+ivan-ottinger@users.noreply.github.com>

---------

Co-authored-by: daledupreez <dale@automattic.com>
Co-authored-by: Ivan Ottinger <25105483+ivan-ottinger@users.noreply.github.com>