Convert E2E infrastructure to ESM (#33846)
Playwright uses a Nodejs loader hook, and the combination of that hook
and changes in Node 20 results in several bugs when it comes to loading
CommonJS files. It's not even clear whether the bugs are in Playwright
or Node.

On the other hand, we can avoid all the bugs (whoever they belong to) by
converting everything over to ESM syntax. So let's do that.