Code Modernization: Replace strpos() with str_starts_with()
Fixes #32854

str_starts_with() was introduced in PHP 8.0 to perform a case-sensitive check indicating if the string to search in (haystack) begins with the given substring (needle).

WordPress core includes a polyfill for str_starts_with() on PHP < 8.0 as of WordPress 5.9.

This commit replaces usage of strpos() with str_starts_with(), making the code more readable and consistent, as well as improving performance in PHP 8+ environments.

While strpos() is slightly faster than the polyfill on PHP < 8.0, str_starts_with() is noticeably faster on PHP 8.0+, as it is optimized to avoid unnecessarily searching along the whole haystack if it does not find the needle.

Note: this doesn't update uses of strpos() in a few places:
- The autoloader since it can be used outside of WordPress.
- The VaultPress plugin since it can still be used on older versions of WordPress, where the str_starts_with() shim did not exist.

Matching core change: https://core.trac.wordpress.org/ticket/58012